TPASPKT-522 Rename DS_STATE_PROPOSED_FOR_DIGITAL_PRESERVATION to DS_STATE_GENERATING_METADATA
Until now, metadata generation was started when the user "proposed" the
dataset for digital preservation. When the dataset was proposed for
preservation, the preservation state of the dataset was then changed to
DS_STATE_PROPOSED_FOR_DIGITAL_PRESERVATION. When metadata generation was
done, the preservation state was changed to
DS_STATE_TECHNICAL_METADATA_GENERATED. So the meaning of
DS_STATE_PROPOSED_FOR_DIGITAL_PRESERVATION has in practice always
been that metadata is being generated.

When preservation process is simplified in TPASPKT-522, the meaning
word "propose" is changed. "Proposing" now means that the user sends the
dataset to be accepted/rejected, so it happens later in the process,
after metada generation. Therefore,
DS_STATE_PROPOSED_FOR_DIGITAL_PRESERVATION would be a misleading name
for state that still means that metadata is being generated.