TPASPKT-1121 Skip HTTP request for empty parameter
If an empty list of file identifiers is provided for
`get_file2dataset_dict`, return an empty dict immediately. Sending a
request with an empty list causes Metax V2 to return a 400 Bad Request
which is not handled by metax-access and causes an exception. We _could_
handle it, but we obviously know that the return value will be empty as
well making the HTTP request pointless in this situation.

If `file_ids` is empty return an empty dict immediately without
performing a HTTP request.