TPASPKT-1397 Fix 'storage_identifier' semantics
`storage_identifier` in V3 migration code was used to refer to the
storage service in several places. This is incorrect; it refers to the
external identifier for a file, which is only unique per storage service
and not in Metax. The Metax file identifier and external file identifier
are `id` and `identifier` in Metax V2, and `id` and `storage_identifier`
in V3.

This confusion was caused by our own `storage_identifier` field name
which was used as an alias for `file_storage.identifier` in several
places such as upload-rest-api and siptools-research.

Fix this issue by normalizing V3 field `storage_service` properly and
changing our internal misleading `storage_identifier` field name to
`storage_service`. This will require a few renames in siptools-research
but should be trivial.