Move states and errors to separate modules
Class types and Metax constants were passed around as method parameters,
possibly as a way to circumvent circular imports.

Move the problematic constants and exceptions to their own modules, as
this makes code more readable and reduces repetition. Note
that we import the exceptions and states also in `metax_access.metax`
even if the module itself doesn't use them; this is done for the sake of
dependent projects such as `upload-rest-api` which import from that
specific location.