TPASPKT-1295 Add 'get_dataset_directory'
Add 'get_dataset_directory' that retrieves the directories and files for
the given directory in a dataset. This is used in the dataset file
browser in fddps-frontend.

The previous 'get_project_directory' technically supported this via the
optional 'dataset_identifier'. Since we only have one specific use case,
it's better to implement a separate function that does the one thing we
need.